+++ boards/riscv/hifive1_revb/Kconfig.board
@@ -3,4 +3,4 @@
 
 config BOARD_HIFIVE1_REVB
 	bool "HiFive1 Rev B target"
-	depends on SOC_RISCV_SIFIVE_FREEDOM
+	depends on SOC_SIFIVE_FREEDOM_E340
