+++ boards/riscv/hifive1_revb/hifive1_revb_defconfig
@@ -1,5 +1,5 @@
-CONFIG_SOC_SERIES_RISCV_SIFIVE_FREEDOM=y
-CONFIG_SOC_RISCV_SIFIVE_FREEDOM=y
+CONFIG_SOC_SERIES_SIFIVE_FREEDOM_E300=y
+CONFIG_SOC_SIFIVE_FREEDOM_E340=y
 CONFIG_BOARD_HIFIVE1_REVB=y
 CONFIG_GPIO=y
 CONFIG_PINCTRL=y
