+++ boards/riscv/hifive_unleashed/Kconfig.board
@@ -3,4 +3,4 @@
 
 config BOARD_HIFIVE_UNLEASHED
 	bool "HiFive Unleashed target"
-	depends on SOC_RISCV_SIFIVE_FU540
+	depends on SOC_SIFIVE_FREEDOM_U540
