+++ boards/riscv/hifive_unleashed/hifive_unleashed_defconfig
@@ -1,5 +1,5 @@
-CONFIG_SOC_SERIES_RISCV_SIFIVE_FREEDOM=y
-CONFIG_SOC_RISCV_SIFIVE_FU540=y
+CONFIG_SOC_SERIES_SIFIVE_FREEDOM_U500=y
+CONFIG_SOC_SIFIVE_FREEDOM_U540=y
 CONFIG_BOARD_HIFIVE_UNLEASHED=y
 CONFIG_CONSOLE=y
 CONFIG_GPIO=y
