+++ boards/riscv/hifive_unmatched/hifive_unmatched_defconfig
@@ -1,5 +1,5 @@
-CONFIG_SOC_SERIES_RISCV_SIFIVE_FREEDOM=y
-CONFIG_SOC_RISCV_SIFIVE_FU740=y
+CONFIG_SOC_SERIES_SIFIVE_FREEDOM_U700=y
+CONFIG_SOC_SIFIVE_FREEDOM_U740=y
 CONFIG_BOARD_HIFIVE_UNMATCHED=y
 CONFIG_CONSOLE=y
 CONFIG_SERIAL=y
