+++ boards/riscv/qemu_riscv32/Kconfig.board
@@ -22,7 +22,7 @@ config BOARD_QEMU_RISCV32_SMP
 
 config BOARD_QEMU_RISCV32_XIP
 	bool "QEMU RISCV32 XIP target"
-	depends on SOC_RISCV_SIFIVE_FREEDOM
+	depends on SOC_SIFIVE_FREEDOM_E340
 	select QEMU_TARGET
 	select HAS_COVERAGE_SUPPORT
 	select CPU_HAS_FPU
