+++ boards/riscv/qemu_riscv32/qemu_riscv32_xip_defconfig
@@ -1,7 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
-CONFIG_SOC_SERIES_RISCV_SIFIVE_FREEDOM=y
-CONFIG_SOC_RISCV_SIFIVE_FREEDOM=y
+CONFIG_SOC_SERIES_SIFIVE_FREEDOM_E300=y
+CONFIG_SOC_SIFIVE_FREEDOM_E340=y
 CONFIG_BOARD_QEMU_RISCV32_XIP=y
 CONFIG_CONSOLE=y
 CONFIG_SERIAL=y
