+++ boards/riscv/sparkfun_red_v_things_plus/sparkfun_red_v_things_plus_defconfig
@@ -1,8 +1,8 @@
 # Copyright (c) 2022 TOKITA Hiroshi <tokita.hiroshi@fujitsu.com>
 # SPDX-License-Identifier: Apache-2.0
 
-CONFIG_SOC_SERIES_RISCV_SIFIVE_FREEDOM=y
-CONFIG_SOC_RISCV_SIFIVE_FREEDOM=y
+CONFIG_SOC_SERIES_SIFIVE_FREEDOM_E300=y
+CONFIG_SOC_SIFIVE_FREEDOM_E340=y
 CONFIG_BOARD_SPARKFUN_RED_V_THINGS_PLUS=y
 CONFIG_GPIO=y
 CONFIG_PINCTRL=y
