--- soc/riscv/riscv-privileged/sifive-freedom/Kconfig.series
@@ -1,11 +1,0 @@
-# RISCV_SIFIVE_FREEDOM SOC implementation
-
-# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
-# SPDX-License-Identifier: Apache-2.0
-
-config SOC_SERIES_RISCV_SIFIVE_FREEDOM
-	bool "SiFive Freedom SOC implementation"
-	select RISCV
-	select RISCV_PRIVILEGED
-	help
-	  Enable support for SiFive Freedom SOC
