+++ soc/riscv/sifive_freedom/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Copyright (c) 2024 Nordic Semiconductor
+# SPDX-License-Identifier: Apache-2.0
+
+add_subdirectory(common)
+add_subdirectory(${SOC_SERIES})
