+++ soc/riscv/sifive_freedom/Kconfig.soc
@@ -1,0 +1,4 @@
+# Copyright (c) 2024 Nordic Semiconductor ASA
+# SPDX-License-Identifier: Apache-2.0
+
+source "soc/riscv/sifive_freedom/*/Kconfig.series"
