+++ soc/riscv/sifive_freedom/e300/Kconfig.defconfig.e340
@@ -1,0 +1,5 @@
+# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC
+	default "e340" if SOC_SIFIVE_FREEDOM_E340
