+++ soc/riscv/sifive_freedom/e300/Kconfig.soc
@@ -1,0 +1,20 @@
+# RISCV_SIFIVE_FREEDOM SOC configuration options
+
+# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
+# SPDX-License-Identifier: Apache-2.0
+
+choice
+	prompt "SiFive Freedom SOC implementation"
+	depends on SOC_SERIES_SIFIVE_FREEDOM_E300
+
+config SOC_SIFIVE_FREEDOM_E340
+	bool "SiFive Freedom SOC implementation"
+	select ATOMIC_OPERATIONS_C
+	select INCLUDE_RESET_VECTOR
+	select RISCV_ISA_RV32I
+	select RISCV_ISA_EXT_M
+	select RISCV_ISA_EXT_A
+	select RISCV_ISA_EXT_ZICSR
+	select RISCV_ISA_EXT_ZIFENCEI
+
+endchoice
