+++ soc/riscv/sifive_freedom/u500/Kconfig.series
@@ -1,0 +1,13 @@
+# RISCV_SIFIVE_FREEDOM SOC implementation
+
+# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC_SERIES_SIFIVE_FREEDOM_U500
+	bool "SiFive Freedom U500 SOC implementation"
+	select RISCV
+	select RISCV_PRIVILEGED
+	select RISCV_PRIVILEGED_STANDALONE
+	select SOC_FAMILY_SIFIVE_FREEDOM
+	help
+	  Enable support for SiFive Freedom U500 SOC
