+++ soc/riscv/sifive_freedom/u500/clock.c
@@ -7,7 +7,7 @@
 #include <zephyr/init.h>
 #include <zephyr/devicetree.h>
 #include <zephyr/sys/util.h>
-#include "fu540_prci.h"
+#include "prci.h"
 
 BUILD_ASSERT(MHZ(1000) == DT_PROP(DT_NODELABEL(coreclk), clock_frequency),
 	"Unsupported CORECLK frequency");
