+++ soc/riscv/sifive_freedom/u700/Kconfig.defconfig.series
@@ -1,9 +1,10 @@
+# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
 # SPDX-License-Identifier: Apache-2.0
 
-if SOC_SERIES_RISCV_SIFIVE_FREEDOM
+if SOC_SERIES_SIFIVE_FREEDOM_U700
 
 config SOC_SERIES
-	default "sifive-freedom"
+	default "u700"
 
 config SYS_CLOCK_HW_CYCLES_PER_SEC
 	default 32768
@@ -32,4 +33,6 @@ config MAX_IRQ_PER_AGGREGATOR
 config NUM_IRQS
 	default 64
 
-endif # SOC_SERIES_RISCV_SIFIVE_FREEDOM
+source "soc/riscv/sifive_freedom/u700/Kconfig.defconfig.u*"
+
+endif # SOC_SERIES_SIFIVE_FREEDOM_U700
