+++ soc/riscv/sifive_freedom/u700/Kconfig.defconfig.u740
@@ -1,0 +1,5 @@
+# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC
+	default "u740" if SOC_SIFIVE_FREEDOM_U740
