+++ soc/riscv/sifive_freedom/u700/Kconfig.series
@@ -1,0 +1,13 @@
+# RISCV_SIFIVE_FREEDOM SOC implementation
+
+# Copyright (c) 2017 Jean-Paul Etienne <fractalclone@gmail.com>
+# SPDX-License-Identifier: Apache-2.0
+
+config SOC_SERIES_SIFIVE_FREEDOM_U700
+	bool "SiFive Freedom SOC U700 implementation"
+	select RISCV
+	select RISCV_PRIVILEGED
+	select RISCV_PRIVILEGED_STANDALONE
+	select SOC_FAMILY_SIFIVE_FREEDOM
+	help
+	  Enable support for SiFive Freedom U700 SOC
